Support exchangefn=NULL in gasneti_sharedLimit
This commit updates logic and comments in `gasneti_sharedLimit()` to
accept a `NULL` value for the `exchangefn` argument.  Doing so will
utilize a new AM-based `gasneti_host_sumu64()`, which not only
eliminates unnecessary off-host communication, but additionally
performs the needed reduce-to-all directly, rather than performing
a gather-to-all followed by a summation.

Once deployed, this is anticipated to provide a significant reduction
in off-host communication for this step with a corresponding reduction
in startup time at scale.  The exception is smp-conduit, where there
is never any off-host communication but their is still an expected
reduction in the communication volume in this step.