Support barrierfn=NULL in gasneti_sharedLimit
This commit updates logic and comments in `gasneti_sharedLimit()` to
accept a `NULL` value for the `barrierfn` argument.  Doing so will
utilize the AM-based `gasneti_host_barrier()`, which was originally
implemented as part of the solution to bug 3480.

Once deployed, this is anticipated to provide a significant reduction
in off-host communication for this step with a corresponding reduction
in startup time at scale.  The exception is smp-conduit, where there
is never any off-host communication and thus no appreciable advantage.