ofi: prioritize Cray PMI over MPI as a spawner
This commit ensures that in the absence of a user-provided setting for
`GASNET_SPAWN_CONTROL`, use of Cray PMI for spawning of ofi-conduit
applications will be given priority over MPI.

This reverses the default ordering in which use of `srun` or `mpirun`
(rather than `gasnet_run[conduit]`) normally assumes MPI should be
used.  This is motivated by an observation that some versions of HPE
Cray MPI set `FI_CXI_*` environment variables in to values which
conflict with what ofi-conduit wants for reliable AMs.