Update gasneti_get_dissem_peers*() to use const
This commit updates `gasneti_get_dissem_peers()`,
`gasneti_get_dissem_peers_pshm()` and their callers to utilize a
vector of `const gex_Rank_t` elements.  This change, along with an
expansion of comments, is intended to enforce and clarify that the
vector result is owned by the implementation and is not for the caller
to modify or free.