ofi: use AM-based bootstrap collectives
This commit follows-up on pull request #542 by implementing dynamic
dispatch of the Barrier and Exchange bootstrap collective operations
in ofi-conduit.  This uses the same approach as in ibv- and
aries-conduits.  This approach allows use of a single identifier for
all instances of the corresponding operation, with the selection of
the spawner-based or AM-based being made dynamically based on the
value of `gasneti_attach_done`.

Prior to this commit, the spawner-based Barrier and Exchanges were
used exclusively.  Therefore, this work will improve startup times
and/or memory usage at large scale by reducing dependence on the
spawner-based implementations.

Additionally, the one use of `gasneti_spawner->Cleanup()` has been
replaced by `gasneti_bootstrapCleanup()`, which predates this commit.
If nothing else, this provides uniformity.