ofi: implement optional GEX_TI_*
This commit implements `GEX_TI_{ENTRY,IS_REQ,IS_LONG}` in
ofi-conduit's `gex_Token_Info()`.

The implementation of `GEX_TI_IS_REQ` allocates an otherwise unused
bit in the on-the-wire header (which is used in-place as the token).
In addition to its use for this query, this bit is now also used to
assert that requests and replies only arrive on the proper completion
queues.

Additionally, this commit adds comments on the layout of the AM header
struct, and makes its (previously implicit) ABI padding explicit.