testalltoall: add new options, changing defaults
This commit adds `-min-size` and `-step-size` options to
`testalltoall` to allow control over the size of AM Medium payloads.
It is now possible to request that payload sizes vary over a given
range, either in fixed steps or uniformly randomly.

This commit adds `-no-reply` and `-reply` options to control whether
the message flow is one-way or round-trip, respectively.  Previously
the flow was always round-trip.

Perhaps most significantly, this commit makes a one-way flow with
uniformly distributed random payload sizes the defaults.  This is
consistent with the default random ordering of peers and makes this
test's default behavior a better proxy for the patterns seen in MHM2.

In addition to describing the new options in the usage output, the
usage output now documents the default for the preexisting options.

Finally, test output now reports all of the option settings.