ofi: opt-in work-around for bugs 4179 and 4461
This commit implements and documents a special environment setting
`GASNET_OFI_RECEIVE_BUFF_SIZE=single` (case insensitive) as a means to
request that the multi-receive buffers be sized to each hold only a
single received AM.  This has been observed to be an effective
work-around for bugs 4179 and 4461 (see brief descriptions below).
However, there is a measurable performance penalty for use of this
work-around

The new logic automatically adjusts the buffer size for the value of
`GASNET_OFI_MAX_MEDIUM`, as well as enlarging the default value of
`GASNET_OFI_NUM_RECEIVE_BUFFS`.  This makes this much simpler (and
safer) to deploy than manually setting both variables.

The logic to compute the largest AM size has moved earlier in the
initialization process, and been made marginally more robust by
removing the assumption that the header for a packed Long cannot be
smaller than for a Medium.

Finally, a bounds check for the minimum permissible value of
`GASNET_OFI_RECEIVE_BUFF_SIZE` has been corrected along with its error
message.

Issues addressed by this commit:

+ In bug 4179, the verbs provider in libfabric 1.11 and older may
experience corruption of AMs.

+ In bug 4461, certain AM traffic patterns lead to misbehavior with the
cxi provider.

In both cases, this commit works-around what is believe to be a defect
in libfabric and/or one of its providers.