ofi: cleanup/factor RDMA injection logic
This commit revises the macros `OFI_WRITE` and `OFI_READ`, and
analogous logic in the injection of non-bulk puts and Long payloads.
This revision is centered around use of new inline functions for
determination of the address and keys used for RDMA injection.  These
functions should be more maintainable than the macros they replace and
in particular provide a better environment for implementation of
multi-endpoint support.

Additionally, `OFI_WRITE` and `OFI_READ` (which are still macros) are
now implemented in terms of a common macro, eliminating near-verbatim
duplication.