ofi: cleanup looping in Long payload injection
This commit replaces two `while` loops with appropriate macros in the
logic to launch an RDMA Put for the payload of a (not packed) AM Long.

This improves the readability and maintainability of the code, and in the
case of the stall for RDMA completion, ensures that the loop honors
`gasnet_set_waitmode()`.