ofi: split GASNET_OFI_INJECT_LIMIT
This commit recognizes the possibility that `FI_MSG` and `FI_RMA`
endpoints may have different `tx_attr->inject_size`.

Previously, the value of the `FI_MSG` endpoints was used as the
default value for the `GASNET_OFI_INJECT_LIMIT` environment variable,
which set the variable `max_buffered_send`.  However, that variable
was used not only for `send` but also for `write`.  In particular, for
RMA put with `lc_opt != GEX_EVENT_DEFER`.

The primary changes are:

+ The `tx_attr->inject_size` for the `FI_MSG` endpoints is still used
  as the default value for `max_buffered_send`.  However the env var
  is now named `GASNET_OFI_AM_INJECT_LIMIT', with the prior name also
  accepted if the new variable is not set.
+ The `tx_attr->inject_size` for the `FI_RMA` endpoint is now used as
  the default value for a new `max_buffered_write`, which governs
  the use if `FI_INJECT` for `write`s.  The corresponding new env
  var is names `GASNET_OFI_RMA_INJECT_LIMIT`.

Additionally, rather than leaking a single `struct fi_info`, the
conduit now leaks two.  However, they are now stored in global
variables (rather than on the stack) in case they are valuable in the
future.