ofi: pass (tm,rank) to internal RMA functions
This commit changes the signature of the `gasnetc_rdma_*()` functions
to take the `(tm,rank)` tuple in place of the jobrank.

There are no "real" changes in this commit because the `(tm,rank)`
tuple is converted to a `jobrank` local variable immediately upon
entry to the modified functions.  Later commits will make more
meaningful use of the tuple.