ofi: shift registration bookkeeping to EP
This commit shifts the bookkeeping for memory registration from a
conduit-specific Segment to a conduit-specific EP.  The former data
structure is no longer needed and it has been removed.  However, the
hook on `gex_Segment_Destroy()` remains to provide an implicit
"unbind" when one destroys a segment bound to an idle endpoint.

This change ensures that binding of the same segment to multiple
endpoints will be handled correctly, as soon as multiple endpoints are
supported.