ofi: avoid pedantic warnings checking callbacks
This commit factors a "check for the right callback function" idiom
in order to express it in a way which does not trigger pedantic
warnings from (at least) gcc.  The problem with the prior use of
`gasneti_assert_ptr()` is that conversion from pointer-to-function
to `void*` is eliciting such warnings even in the presence of an
explicit cast.  Only adding an intermediate cast to a non-pointer
type has been found effective.  The new macro hides this ugliness.