coll: disaggregate p2p allocation
This commit replaces the monolithic allocation of the "p2p" structure
and three subordinate arrays with four distinct allocations, which may
now be of variable size (including absent entirely).

This commit adds dynamic p2p field allocation to several AM handlers
in the reference collectives, which are used by a wide range of
collective algorithms. They differ from the prior implementation in
that they will dynamically allocate the potentially large p2p fields
(`state`, `counter` and `data`) if they have not been allocated
previously, realloc'ing if an earlier allocation was too small.  Such
dynamic allocations are as large as needed to encompass the AM
Request's specific operation, subject to a minimum growth to mitigate
worst-case behavior.

Unfortunately, it is not practical to allocate the exact size for the
full operation "on first touch", since providing sufficient
information to an AM handler would require up to three additional
arguments and non-trivial logic changes in the multiple requesting
code paths.

The AM handlers which manipulate p2p structures have also been updated
to hold a lock to prevent concurrent realloc check and in debug builds
they check for possible overruns when releasing that lock.

While this commit does not introduce any logic to prevent the
collective algorithms from allocating the worst-case sizes, the
expectation is that in the future (at least for scalable collectives
like broadcast and reduce) the "right sized"-sized allocation will
take place in the first execution of the collective poll function,
eliminating the need for realloc for the (hopefully common case) that
local initiation occurs prior to an AM arrival.

Additionally, the "right sized" allocation will "finalize" the p2p
struct (preventing any further realloc in a debug build) and thus
eliminating the need to hold a lock in the poll function.

In the collective algorithms in which this work is intended to be
used, the worst case for dynamic allocations will be
`O(log(team_size))` in count.  The size of `state` and `counter`
allocations are also expected to be logarithmic in team size -- a
dramatic reduction from the prior linear scaling.  The `data`
allocation may still be as large as the maximum eager buffer size but
populated in a number of operations logarithmic in team size.

While the logic continues to place the p2p structures back on a
freelist for quick reuse, the three (potentially large) subordinate
arrays are now freed.  This has the effect of adding as many as three
allocation and three frees to the lifecycle of every collective
operation.  However, the prior behavior was to have these potentially
large allocations leaked permanently.  This change in particular,
addresses a portion of the scalability problems documented in bug
4903.