coll: use scalable storage in reductions
This commit follows up on the introduction of dynamic allocation of
p2p fields within collectives AM Request handlers with modifications
to the four imp four implementations of reduce (which include both
reduce-to-one and reduce-to-all operations) to utilize this
capability.

In all cases the first run of the collective poll functions for a
reduction now allocate the entire necessary sizes (if any) of the p2p
fields `state`, `counter` and `data` and "finalize" the p2p.