fixup - ofi: initial memory kinds support
This commit makes a one-line change to ensure that compilation of the
caller and callee for the `GASNETC_MK_CREATE_HOOK` are conditional
on the same preprocessor expression.

This resolves `undefined reference to 'gasnetc_mk_create_hook'` errors
seen when linking with a non-kinds build of ofi-conduit using the PGI
or NVHPC compiler families.

Even though the prior difference between the controlling preprocessor
expressions lead to a caller without a callee, this discrepancy went
unnoticed in PR testing.  Experiments using `#warning` for testing
reachability and `nm` for examining code generation support a
hypothesis that all other supported compiler families we use on Dirac
are *discarding* the call.  Indeed, in a non-kinds build all inputs
lead to one of `gasneti_fatalerror()`, `GASNETI_RETURN_ERRR()` or
`gasneti_unreachable_error()` prior to calling the hook.