GASNet CI: test annotations for bug 4151
This commit provides conditional setting of the environment variable
needed to work around bug 4151 on impacted systems.  This is intended
to replace blanket setting for the entire harness run, which is known
to degrade the performance of host-memory RMA operations.  This is
better aligned with our documentation, which advises against setting
this variable unless needed.