UPCXX CI: test annotations for bugs 4151 and 4504
This commit provides conditional setting of the environment variable
needed to work around bug 4151 on impacted systems, and the compile
flag needed to work-around bug 4504 (disabling part of the test).

The annotations for bug 4151 are intended to replace simply setting
the environment variable for the entire harness run, since it is known
to degrade the performance of host-memory RMA operations.  This is
better aligned with the GASNet-EX documentation, which advises against
setting this variable unless needed.