configure: revise handling of force arch
This commit replaces an undocumented family of `--enable-arch-FOO`
options with `--enable-force-arch=FOO`.  This ensures these options
are mutually exclusive, as well as hopefully conveying (via the word
`force`) that these override the defaults.  This includes addition of
a `generic` value for the new option, to prevent the probes which
would identify one of the other valid values by default.