bug 4510: Fix a livelock in testinternal/rwlock
The test was previously making unbounded calls to `gasneti_rwlock_try{rd,wr}lock()`,
which is not guaranteed to ever succeed (even with a fair rw mutex
implementation). This was leading to intermittent slowdowns and hangs on
at least one platform.

Convert the test to instead perform a bounded number of trylock
operations, and then fallback to a blocking lock.