coll: add a missing RMB to resolve bug 4462
This commit adds a `gasneti_sync_reads()` missing from the Bruck's
dissemination implementation of GatherAll.

As reported in bug 4462, prior to this change we have seen occasional
failures of `testcoll-par -p 1 200 1` on Summit, where the IBM Power9
CPU has a more relaxed memory model than the x86/x86_64 CPUs which
dominate our testing and development.

With this change, no failures were observed in 1000 trials.  Compare
that to prior to this commit when no more than 40 consecutive
successes were achieved in any of five attempts to reach 100.