ibv: fix debug output for INLINE-sent AMs
When AMs are small enough, they may be sent using an on-stack buffer
and `IBV_SEND_INLINE`.  In this case, the logic which prints various
information in an error path does not have access to the original
buffer containing the `args` or `nybytes` fields.  So, this commit
repairs the logic to avoid a `NULL` deference in this error path.