Implement GASNET_USE_HUGEPAGES (bug 4533)
This commit implements and documents a new `GASNET_USE_HUGEPAGES`
environment variable.  The default value of `1` corresponds to the
only behavior available prior to this commit: it enables the use of
`libhugetlbfs` for certain allocations, in builds which have been
configured to use it.  The non-default `0` forces GASNet-EX to instead
allocate normal pages, much as it would if `libhugetlbfs` not been
enabled at configure time.

Note that this environment variable only controls allocation calls in
GASNet-EX.  It notably does not reverse/disable/override any behaviors
which come from linking and running with `libhugetlbfs` and
environment variables such as those set in a `craype-hugepages`
environment module.

This resolves bug 4533: "RFE: run-time disable for hugepage support".