Smarter default for GASNET_USE_HUGEPAGES
This commit updates the default value of `GASNET_USE_HUGEPAGES` to be
"1" if `HUGETLB_DEFAULT_PAGE_SIZE` is set, and "0" otherwise.  In the
case of a Cray XC or HPE Cray EX system, this means defaulting to use
of hugepages only when a `craype-hugepages*` environment module is
loaded.

(cherry picked from commit 6c0dc7689ea19f192fbec7a20583a4e15dd86c64)

Conflicts:
	ChangeLog (update ommitted)