ofi: improved support for recent cxi provider
The changes in PR#571 introduce no performance degradation for the
older libfabric/1.15.0.0, but for libfabric/1.15.2.0 it adds a table
lookup (for the provider-allocated memory registration key) in the
critical path of every communications path.  This commit removes that
table lookup, resolving bug 4561 "cxi provider performance
regression".

The changes in this commit are motivated by the observation that even
though the cxi provider in libfabric/1.15.2.0 was the first to
preserve the `FI_MR_PROV_KEY` bit in `mr_mode` when calling
`fi_getinfo()`, it does not *require* the client to set that bit.  So,
we can statically remove `FI_MR_PROV_KEY` support from the library at
compile time for *all* known versions of the cxi provider.

In addition to changing the configure defaults for the cxi provider,
this commit makes changes in the run time behavior for the "generic"
provider to arrive at the same behavior for the cxi provider.  In
particular, handling of the `FI_MR_VIRT_ADDR` and `FI_MR_MR_PROV_KEY`
bits is updated and extended to include preference in the "generic"
case for *clearing* the `FI_MR_MR_PROV_KEY` bit.