fixup: ofi: add ident strings for provider feat...
The work in PR#571 to add ident strings for some provider properties
changed `GASNETC_OFI_HAS_MR_SCALABLE` from an integer expression to a
Boolean one (except in the "generic provider" case).  That change
resulted in compiler warnings from a `switch` statement (exclusive to
libfabric versions prior to 1.5) which includes `-1` as one of its
cases.

This commit replaces the problematic `switch` with logic free from the
warnings (and more readable IMO).