ofi: decouple FI_MR_ALLOCATED from FI_MR_VIRT_ADDR
This commit removes the assumption that the `FI_MR_ALLOCATED` in
`mr_mode` follows the `FI_MR_VIRT_ADDR` bit.  This allows for
the possibility of supporting `--enable-segment-everything` with
a (hypothetical) provider which asserted `FI_MR_VIRT_ADDR` but not
`FI_MR_ALLOCATED`.

This completes efforts to shed the assumptions about `mr_mode` bits
which had arisen with the original "BASIC" and "SCALABLE" modes.