ofi: return gni provider defaults to "FI_MR_BASIC"
This commit reverts a recent change of default memory registration
mode for gni provider, setting it back to `FI_MR_BASIC`.  This is in
response to the findings in bug 4565.

While this commit corrects the behavior for `--with-ofi-provider=gni`
(or equivalent configure-time auto detection), the case of
`--with-ofi-provider=generic` remains broken.

This reverts commit d89529501318de720b3c4b00136c56c7c6ae6db9.