Revise RDMADISSEM barrier to use four phases
This commit changes the allocation indexing of the "inboxes" used by
the `RDAMDISSEM` barrier to use four phases, where two were used
before.  However, the most important change is to how (or rather
"what") is reset upon completion of a barrier step.  Where previously
we would overwrite the inbox which had just matched, we now overwrite
an older one which is provably "idle".  By that I mean that, under
some reasonable assumptions, we know that any/all agents which might
write to that space have completed doing so by the virtue of having
completed two intervening barriers.

This commit resolves the root cause of bug 4427 and, in theory, the
`RDMADISSEM` portion of bug 4542.