Docs updates for hwloc-related RFEs 4170 and 4171
This commit updates documentation for work done to resolve
hwloc-related RFEs 4170 and 4171.

The top-level `README` now contains documentation on the use of the
`_TYPE` variables, including the new capabilities (ranks and
arithmetic).  There are corresponding reductions in the `README` files
of ibv- and ofi-conduit, which now reference the top-level one.

Appropriate entries have been added to the `ChangeLog`.