fixup - hwloc: "Rank" support (bug 4171)
Commit b9c56e74 in PR#581 caused a build regression for ibv- and
ofi-conduits (both of which call `gasneti_getenv_hwloc_withdefault()`)
on systems without `hwloc` located at configure time:

```
[...]/gasnet/other/hwloc/gasnet_hwloc.c:
In function 'gasneti_getenv_hwloc_withdefault':
[...]/gasnet/other/hwloc/gasnet_hwloc.c:335:7:
error: label 'try_suffix' used but not defined
  335 |       goto try_suffix;
      |       ^~~~
```

This commit resolves the regression by relocating a block of
approximately 30 lines of code in `gasneti_getenv_hwloc_withdefault()`
to lie outside the preprocessor conditionals for hwloc support.  The
block in question already began with a label (the one in the error
message) and ended with a `goto`.  So the only changes other than pure
movement of code are (1) one added `goto` to the label which starts
the block and (2) removal of the `goto` which previously ended the
block (but became a no-op in the new location).