fixup - Implement gasneti_str[n]casecmp() for bug 4582
The POSIX specifications for `strcmp()` `strcasecmp()` and
`strncasecmp()` are silent as to the behavior when either string
argument is `NULL`.  To the best of my understanding, this means the
behavior is undefined (literally: no normative text defines the
behavior).  So, when implementing `gasneti_strcasecmp()` I made the
assumption that existing code which would adopt the new function
*should not* be passing `NULL` to `strcmp()` and that is was therefore
appropriate to assert that both string arguments were non-`NULL`.

However, with merge of PR#584 to make use of `gasneti_strcasecmp()` in
environment parsing, these new non-`NULL` assertions have been seen to
fail.  Looking deeper, I've found that in (at least) one case the use
of `gasneti_strdup(NULL)` to allocate a temporary string used for
case normalization was the only thing preventing us from passing
`NULL` to `strcmp()`.  This is a subtly I had overlooked.

This commit modifies the behavior of `gasneti_str[n]casecmp()` to
treat a `NULL` argument as indistinguishable from an empty string and
documents this *defined* behavior in comments.  This is believed to be
more reliable than requiring an explicit check in callers, and there
is no reason to believe the additional logic is a realistic
performance concern.