ofi: per-provider minimum versions
This commit makes interrelated changes to address multiple bugs filed
against ofi-conduit with respect to the minimum versions of libfabric
required by some providers.

1. This commit documents 1.15.0 as the  minimum libfabric version for
"tcp" provider.  This resolves bug 4579: "Document tcp provider
floor" by establishing a minimum based on recent testing.

2. This commit documents 1.12.0 as the minimum libfabric version for
"verbs" provider.  This resolves (as WORKSFORME) bug 4179: "ofi:
failures specific to verbs provider".  This resolves (as INVALID) bug
4517: "ofi/verbs 'no associated AM handler function' errors with
libfabric 1.11 or older" by setting the minimum libfabric version for
this provider to one without the defect.

3. In addition to adding new documented floors for two provider, this
commit reorganizes ofi-conduit's `README` to collect the version
floors in a single list allowing for some clean-up to the lists where
such information previously appeared.

4. This commit adds configure-time enforcement of the floor versions
for the four providers which document one.  Of course, this is not
applied to the case of the "generic" provider, since a provider is not
known then.

5. This commit adds run-time enforcement of the four provider-specific
floors.  This serves to ensure enforcement for the "generic" provider
case, and should *normally* be redundant with the configure-time
checks when building for a single provider.  However, the run-time
checks use the version information reported dynamically by the
library, rather then what the headers report statically.  So, an older
library installed on compute nodes may be detected even though the
configure-time check had passed on a front-end node.

Together, items #4 and #5 resolve bug 4569: "RFE: per-provider version
floor enforcement".

The intention of enforced floors is to give users a better experience
than the "bad behaviors" (such as random data corruption) with older
versions which led to the selection of the specific documented floors.
However, there are (intentionally undocumented, other than here) means
to circumvent these checks.  The configure-time check can be disabled
using `--enable-allow-untrusted-ofi-provider`.  The run-time check can
be converted to a console message by setting environment variable
`GASNET_OFI_ALLOW_UNTRUSTED_PROVIDER=1`.