ofi: configure support for --with-ofi-provider=efa
This commit adds support in `configure` for `--with-ofi-provider=efa`.
Specifically, this establishes the provider as one with a memory mode
equivalent to the legacy `FI_MR_BASIC`, supporting `FI_AV_TABLE`, and
requiring the work-around for bug 4376 (infrequent `FI_EAGAIN` from
`fi_recvmsg()`).  These are based on testing of both GASNet's in-repo
tests and `make dev-check` in UPC++.

Currently the ofi-conduit `README` remains silent with respect to this
provider.