fixup - ofi: per-provider minimum versions
This commit updates the enforcement of minimum libfabric versions
for the "verbs" provider to allow HPE's libfabric 1.11.x.y.z
versions which are installed on HPE Cray EX systems with HPE's
Slingshot-10 NIC.

Experience has shown that the "verbs" provider in HPE's libfabric does
not suffer from most of the defects listed in bug 4179 which motivated
the minimum of 1.12.  The exception to that is bug 4517.