ofi: "generic" provider selection of old "udp"
Prior to libfabric 1.12.0, the "udp" provider was identified as "UDP"
in the results of `ofi_getinfo()`.  Due to case-sensitive comparison
in ofi-conduit, the implementation of the "generic" provider would
fail to match this provider name against `GASNETC_OFI_PROVIDER_LIST`.

This commit uses case-insensitive string comparison to resolve this
issue.