fixup - ofi: per-provider minimum versions
This commit corrects the behavior for a libfabric version too old for
the specified (or detected) provider in the presence of
`--enable-ofi=probe`.  As implemented previously, the failure of the
version check was *always* fatal.  Now only in the case of
`--enable-ofi` does the behavior remain fatal.

Additionally, the logic for validating provider-specific version
floors is relocated to follow the message (if any) about an
auto-detected provider and converted to a CHECKING/RESULT pair.
This makes for a much more logical presentation to the end user
in the case of failures.