Chapel CI: update/correct CHPL_HOST_COMPILER
This commit corrects the logic for selection of `CHPL_HOST_COMPLIER`
on HPE Cray EX (Shasta) systems.  We already "knew" that we are not
supposed to set this variable on Cray XC systems like Cori, and had
been using the `os_cnl` feature for that purpose.  However, that is
not set for Shasta systems because we don't cross-compile.  So, this
commit updates the two "legacy" cases (`cc_intel` and `cc_pgi`) and
adds logic for `cc_clang` to ensure we don't set this variable in the
presence of the corresponding `prgenv-*` harness features (recently
added to pushbuild).