Enable nesting of gasneti_begin_nbi_ff()
This commit adds a depth counter to the implementation of
`gasneti_begin_nbi_ff()` to allow simple nesting of calls.  The only
"catch" is that if the nbi_ff aop is on the iop stack it must be
the top of that stack to call `gasneti_begin_nbi_ff()`.  So a new
`gasneti_nbi_ff_ok()` is added to allow callers to know if they can
safely use `gasneti_begin_nbi_ff()`.