UPCXX CI: Add upcxx-info to upcxx.log
This change will generate messages in the output of the form:

```
Config file not found: .../upcxx-inst/gasnet.debug/config-summary.txt
```

This is normal and expected, because when run under the default UPCXX_GASNET
setting and any other config that uses existing GASNet build tree, UPC++
will not configure GASNet meaning these files are expected to be
missing. Even with UPCXX_GASNET setting where UPC++ configures GASNet,
it uses --enable-single so some of these files (for the other codemode)
are still expected to be missing.