UPCXX CI: suppress new GCC 13.x warnings
Starting with 13.1.0, g++ is now warning about our
explicit/intentional self-move operations.  So, this commit adds
`-Wno-self-move` to the flags when compiling issue547 with g++, in
addition to clang++.

Note that our floor of g++ 6.4.0 does *not* recognize `-Wself-move`,
but (as with all GCC compilers I am aware of), silently ignores
`-Wno-*` arguments that it does not recognize.