GASNet CI: add tests for non-default trace/stats
This commit adds new tests to the `libgasnet` suite to build and link
a test, with each enabled conduit and both `seq` and `par` thread
modes, with a library configured with non-default ("alternate")
options for tracing and/or statistics collection.  For instance,
enabling one or both in an NDEBUG build, or disabling them in a DEBUG
build.

Similar to the "libgasnet -pedantic" tests, harness features must
be set to opt-in to these new tests.
There are three distinct features:

+ `libgasnet_alt_t`   test with inverted --(en|dis)able-trace
+ `libgasnet_alt_s`   test with inverted --(en|dis)able-stats
+ `libgasnet_alt_ts`  test with inverted --(en|dis)able-trace and --(en|dis)able-stats

This set of tests is intended to reveal errors in use of the
preprocessor tokens associated with trace and stats support.  For
instance, enabling the `libgasnet_alt_ts` in a DEBUG build could
discover an assertion using a variable with a definition guarded by
`GASNETI_STATS_OR_TRACE`.