GASNet CI: update/refine bug 4384 work-around
This commit introduces `has_bug_4384` to replace erroneous use of
`ucx-conduit` (which is never defined) as a harness feature to control
setting of the environment variables which work-around bug 4384.
In addition to being correct and more descriptive, the use of a
feature named for the bug allows for the future possibility of a
UCX release which doesn't have this bug.

In addition to replacing two erroneous `AppEnv` lines, this commit
adds sixteen such lines which were missing from the `gasnet-tests`
suite.  This lack had been masked in pushbuild's `EX-summit-ucx-pgi`
by harness-wide setting of the work-around variables, and in
`EX-dirac-ucx-gcc` by bug 4640 which had prevented running the
impacted tests.