ucx: selective enable of native atomics
This commit resolves bug 4594: "UCX should not enable native atomics
unconditionally" by cloning the logic for native-vs-AM atomics
selection from aries-conduit.  This logic gives precedence to the
user's `GEX_FLAG_AD_FAVOR_*` flags (if any), and otherwise selects the
AM-based solution for the case of a single process and for a single
neighborhood with a "tools-safe" datatype.