ucx: fix exit-time crashes on Summit
This commit resolves bug 4655 "ucx: bad exits on Summit" by reordering
the finalization calls to the UCX library and the spawner to properly
mirror their order of initialization.

W/o this change, `make run-tests-seq` on two nodes of Summit using
mpi-spawner will crash in 41 of 43 tests (only single-process
testtools and testmisc do not crash).  It is believed that this is due
to some hidden interaction between finalization of IBM Spectrum MPI
and UCX's use of the same libibverbs used by MPI.  However, no
specific cause was identified.