fixup - ucx: use AM-based bootstrap collectives
This commit resolves a lingering "TODO" from the initial introduction
of AM-based bootstrap collectives in ucx-conduit.  In doing so, it
resolves a hang of `gasnet_init()` in `GASNET_SEGMENT_EVERYTHING`
builds recently introduced by the resolution of bug 4280.

Prior work to utilize more AM-based collectives for bootstrap
communication did not convert the barrier in ucx-conduit's
`gasnetc_attach_primary()` from spawner-based to AM-based, though the
corresponding change *was* made in all other conduits.

Though not known the time of the prior work, the unchanged
spawner-based barrier worked in a `GASNET_SEGMENT_EVERYTHING` build
*only* due to an undesirable stall in AM injection (subject of bug
4280).  That stall for ucx-level completion of AM sends hid the fact
that UCX might not commit an AM to the network until a later call to
`ucp_worker_progress()` (as via `gasnet_AMPoll()` and its internal
analogues).  With that stall having been recently removed, it became
possible for a client using the legacy init/attach initialization
sequence in an EVERYTHING build to proceed from injecting an AM to
exchange the aux seg bounds to a spawner-based barrier with no
intervening progress.  As a result, at least one other process might
never receive an AM required to complete the exchange, leaving it hung
in `gasnet_init()`

Not only is this change necessary to prevent the hangs described in
the previous paragraph, but it is additionally expected to provide
improved performance for startup at-scale in all configurations.