ucx: comment/clean-up EVERYTHING AM reception
This commit adds or updates numerous comments in the AM reception
logic for the EVERYTHING segment mode, and makes a few minor
"clean-ups" to improve legibility of the code.  Updates the
"future work" in the conduit README as it relates to EVERYTHING.

Additionally appends `// GASNETC_PIN_SEGMENT` comments to several
`#else` and `#endif` directives which are more than "a screen full"
of lines away from the preceding `#if GASNETC_PIN_SEGMENT` or `#else`.

There are no intentional changes to semantics or to runtime behavior.