ucx: don't set the (unused) message tag
Currently ucx-conduit does not make any use of the tag associated with
the UCX-level message APIs.  This commit removes (in favor of `0`)
passing of the source jobrank as message tag, in order to reduce
possible confusion.  In particular, this change helps avoid the
implication that the tag is unavailable for future uses (such as
target-side reassembly of Longs).