tests: add -lbufs option to testlarge
This commit adds an `-lbufs N` option to `testlarge` which changes the
out-of-segment behavior to cycle over `N` buffers, each sized for the
maximum payload.  This is instead of the default behavior in which
a single local buffer is used for every iteration.

The new option is ignored (with a warning) when using in-segment local
addresses.

The name `-lbufs` is chosen to avoid conflict with a possible `-rbufs`
in the future (to cycle over multiple remote buffers).