{Chapel,Legion,UPCXX} CI: support paths and file:/
This commit enhances `external-{chapel,legion,upcxx}/Makefile` to
explicitly handle absolute paths and `file:/` as values of their
respective `URL` variables.

Since `wget` does not support the `file:` protocol, this as an
alternative to requiring `curl` as a pre-requisite for eventual
pushbuild improvements to fetch the various runtime and examples
archives once daily.

In addition to uniform support `wget` and `curl`, the new logic uses a
symlink in place of the copy which `curl` would create.