fixup: {Chapel,Legion,UPCXX} CI: support paths ...
This commit fixes use of the `:` operator in `expr` when the pattern
is a lone `/`.  As documented in an example in the macOS manpage, the
parser may consider a lone `/` to be the division operator.  This
commit uses the recommended work around of prepending a slash to both
the LHS and RHS.

This resolves non-fatal `expr: syntax error` messages seen in nightly
regression testing on macOS, as well as fixing support for full path
values for the URL variables on (at least) macOS.